Added weighted shuffle capability to Order.shuffle
Existing functionality and behaviour is preserved. But a second optional factor parameter, between 0.0 and 1.0 controls the level of randomness applied to the array. A value of 0.0 is pure order. A value of 1.0 (default) is pure randomness with the same result as before. This can be useful when only a small amount of randomness needs to be added to the track order. I actually use it in some of my programs with a factor of 0.05.